Refactor: - Proposal: GhoXLayerActivation_20260203 --> GhoMonadActivation_20260518 - XLayer --> Monad - Add XLayer<>Monad remote lanes / remote lanes test - Constants: add Monad chain
