+++ src/20260518_Multi_GhoMonadActivation/AaveV3Monad_GhoMonadActivation_20260518.sol
@@ -7,10 +7,10 @@ import {GhoCCIPChains} from '../helpers/gho-launch/constants/GhoCCIPChains.sol';
 /**
  * @title Gho X-Layer Launch
  * @author @TokenLogic
- * - Snapshot: https://snapshot.box/#/s:aavedao.eth/proposal/0x251c520f1f1da8287168420fa2d2a73a2eb5342c3c62508553123129dec059b0
- * - Discussion: https://governance.aave.com/t/arfc-launch-gho-on-x-layer-set-aci-as-emissions-manager-for-rewards/23178
+ * - Snapshot: TODO
+ * - Discussion: TODO
  */
-contract AaveV3XLayer_GhoXLayerActivation_20260203 is AaveV3GHOLaunch {
+contract AaveV3XLayer_GhoMonadActivation_20260518 is AaveV3GHOLaunch {
   constructor() AaveV3GHOLaunch(GhoCCIPChains.XLAYER()) {}
   function _setupGhoAaveSteward() internal override {
     // Do not setup Aave Core Steward, will be set up with Gho launch on Aave.
