+++ src/20260518_Multi_GhoMonadActivation/tests/remote-lanes/AaveV3Gnosis_GhoMonadActivation_20260518.t.sol
@@ -4,34 +4,35 @@ pragma solidity ^0.8.0;
 import {AaveV3GHORemoteLaneTest_PreExecution, AaveV3GHORemoteLane_1_6_Test_PostExecution} from '../../../helpers/gho-launch/tests/AaveV3GHORemoteLaneTest.sol';
 import {GhoCCIPChains} from '../../../helpers/gho-launch/constants/GhoCCIPChains.sol';
 import {AaveV3GHOLane} from '../../../helpers/gho-launch/AaveV3GHOLane.sol';
-import {AaveV3Plasma_GhoXLayerActivation_20260203} from '../../remote-lanes/AaveV3Plasma_GhoXLayerActivation_20260203.sol';
-import {AaveV3Plasma} from 'aave-address-book/AaveV3Plasma.sol';
+import {AaveV3Gnosis_GhoMonadActivation_20260518} from '../../remote-lanes/AaveV3Gnosis_GhoMonadActivation_20260518.sol';
+import {AaveV3Gnosis} from 'aave-address-book/AaveV3Gnosis.sol';
 
-uint256 constant PLASMA_BLOCK_NUMBER = 16147000;
+// TODO
+uint256 constant GNOSIS_BLOCK_NUMBER = 00;
 
-contract Plasma_XLayer_AaveV3GHOLane_20260105_Test_PreExecution is
+contract Gnosis_Monad_AaveV3GHOLane_20260105_Test_PreExecution is
   AaveV3GHORemoteLaneTest_PreExecution
 {
   constructor()
     AaveV3GHORemoteLaneTest_PreExecution(
-      GhoCCIPChains.PLASMA(),
-      GhoCCIPChains.XLAYER(),
-      'plasma',
-      PLASMA_BLOCK_NUMBER
+      GhoCCIPChains.GNOSIS(),
+      GhoCCIPChains.MONAD(),
+      'gnosis',
+      GNOSIS_BLOCK_NUMBER
     )
   {}
 
   function _deployAaveV3GHOLaneProposal() internal virtual override returns (AaveV3GHOLane) {
-    return new AaveV3Plasma_GhoXLayerActivation_20260203();
+    return new AaveV3Gnosis_GhoMonadActivation_20260518();
   }
 
   function test_defaultProposalExecution() public virtual {
-    defaultTest('AaveV3Plasma_GhoXLayerActivation_20260203', AaveV3Plasma.POOL, address(proposal));
+    defaultTest('AaveV3Gnosis_GhoMonadActivation_20260518', AaveV3Gnosis.POOL, address(proposal));
   }
 
   function _assertOnAndOffRamps() internal view override {
-    _assertOnRamp_1_6(
-      _localOutboundLaneToEth_1_6(),
+    _assertOnRamp(
+      _localOutboundLaneToEth(),
       LOCAL_CHAIN_SELECTOR,
       ETH_CHAIN_SELECTOR,
       LOCAL_CCIP_ROUTER
@@ -42,8 +43,8 @@ contract Plasma_XLayer_AaveV3GHOLane_20260105_Test_PreExecution is
       REMOTE_CHAIN_SELECTOR,
       LOCAL_CCIP_ROUTER
     );
-    _assertOffRamp_1_6(
-      _localInboundLaneFromEth_1_6(),
+    _assertOffRamp(
+      _localInboundLaneFromEth(),
       ETH_CHAIN_SELECTOR,
       LOCAL_CHAIN_SELECTOR,
       LOCAL_CCIP_ROUTER
@@ -57,25 +58,25 @@ contract Plasma_XLayer_AaveV3GHOLane_20260105_Test_PreExecution is
   }
 }
 
-contract Plasma_XLayer_AaveV3GHOLane_20260105_Test_PostExecution is
+contract Gnosis_Monad_AaveV3GHOLane_20260105_Test_PostExecution is
   AaveV3GHORemoteLane_1_6_Test_PostExecution
 {
   constructor()
     AaveV3GHORemoteLane_1_6_Test_PostExecution(
-      GhoCCIPChains.PLASMA(),
-      GhoCCIPChains.XLAYER(),
-      'plasma',
-      PLASMA_BLOCK_NUMBER
+      GhoCCIPChains.GNOSIS(),
+      GhoCCIPChains.MONAD(),
+      'gnosis',
+      GNOSIS_BLOCK_NUMBER
     )
   {}
 
   function _deployAaveV3GHOLaneProposal() internal virtual override returns (AaveV3GHOLane) {
-    return new AaveV3Plasma_GhoXLayerActivation_20260203();
+    return new AaveV3Gnosis_GhoMonadActivation_20260518();
   }
 
   function _assertOnAndOffRamps() internal view override {
-    _assertOnRamp_1_6(
-      _localOutboundLaneToEth_1_6(),
+    _assertOnRamp(
+      _localOutboundLaneToEth(),
       LOCAL_CHAIN_SELECTOR,
       ETH_CHAIN_SELECTOR,
       LOCAL_CCIP_ROUTER
@@ -86,8 +87,8 @@ contract Plasma_XLayer_AaveV3GHOLane_20260105_Test_PostExecution is
       REMOTE_CHAIN_SELECTOR,
       LOCAL_CCIP_ROUTER
     );
-    _assertOffRamp_1_6(
-      _localInboundLaneFromEth_1_6(),
+    _assertOffRamp(
+      _localInboundLaneFromEth(),
       ETH_CHAIN_SELECTOR,
       LOCAL_CHAIN_SELECTOR,
       LOCAL_CCIP_ROUTER
